airbyte-integrations/connectors/source-databricks/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-databricks"
description = "Source implementation for Databricks SQL."
authors = ["Airbyte <contact@airbyte.io>"]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/databricks"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
[[tool.poetry.packages]]
include = "source_databricks"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^1"
databricks-sql-connector = "^4.0.0"
pandas = "^2.0.0"
pyarrow = "^14.0.0"
numpy = ">=1.21,<2.0"

[tool.poetry.scripts]
source-databricks = "source_databricks.run:run"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.6.1"
requests-mock = "^1.11.0"
freezegun = "^1.4.0"

[tool.poe]
include = [
    "${POE_GIT_DIR}/poe-tasks/poetry-connector-tasks.toml",
]
